Add security reports for Islandora (Drupal).
The neccessary bash scripts have to be copied via Ansible to all
affected servers, then

'bin/security_check_overview.sh check'

can be run to get the report. When it is run

'bin/security_check_overview.sh check mail'

HTML mail version of the report is sent to configured mails (mutt is
supposed to be installed).

See the config variables in the script.